+++ src/azure-cli/azure/cli/command_modules/reservations/tests/latest/test_reservations_commands.py
@@ -2,7 +2,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-
 from azure.cli.testsdk import ScenarioTest
 
 
@@ -39,7 +38,7 @@ class AzureReservationsTests(ScenarioTest):
             'subscription': '00000000-0000-0000-0000-000000000000'
         })
         result = self.cmd('reservations reservation-order-id list --subscription-id {subscription}') \
-                     .get_output_in_json()
+            .get_output_in_json()
         for order_id in result['reservationOrderIds']['value']:
             self.assertIn('/providers/Microsoft.Capacity/reservationorders/', order_id)
 
@@ -55,7 +54,7 @@ class AzureReservationsTests(ScenarioTest):
 
     def test_get_reservation_order(self):
         self.kwargs.update({
-            'reservation_order_id': "0a47417c-cd30-4f67-add6-d631583e09f3"
+            'reservation_order_id': '0a47417c-cd30-4f67-add6-d631583e09f3'
         })
         command = 'reservations reservation-order show --reservation-order-id {reservation_order_id}'
         reservation_order = self.cmd(command).get_output_in_json()
@@ -65,10 +64,10 @@ class AzureReservationsTests(ScenarioTest):
 
     def test_list_reservation(self):
         self.kwargs.update({
-            'reservation_order_id': "0a47417c-cd30-4f67-add6-d631583e09f3"
+            'reservation_order_id': '0a47417c-cd30-4f67-add6-d631583e09f3'
         })
         reservation_list = self.cmd('reservations reservation list --reservation-order-id {reservation_order_id}') \
-                               .get_output_in_json()
+            .get_output_in_json()
         self.assertIsNotNone(reservation_list)
         for reservation in reservation_list:
             self.assertIn(self.kwargs['reservation_order_id'], reservation['name'])
@@ -77,7 +76,7 @@ class AzureReservationsTests(ScenarioTest):
 
     def test_get_reservation(self):
         self.kwargs.update({
-            'reservation_order_id': "0a47417c-cd30-4f67-add6-d631583e09f3",
+            'reservation_order_id': '0a47417c-cd30-4f67-add6-d631583e09f3',
             'reservation_id': 'ae1fbdad-6333-4964-9f4c-83f7e2b7f44f'
         })
         reservation = self.cmd('reservations reservation show  --reservation-order-id {reservation_order_id} '
@@ -89,7 +88,7 @@ class AzureReservationsTests(ScenarioTest):
 
     def test_list_reservation_history(self):
         self.kwargs.update({
-            'reservation_order_id': "0a47417c-cd30-4f67-add6-d631583e09f3",
+            'reservation_order_id': '0a47417c-cd30-4f67-add6-d631583e09f3',
             'reservation_id': 'ae1fbdad-6333-4964-9f4c-83f7e2b7f44f'
         })
         history = self.cmd('reservations reservation list-history --reservation-order-id {reservation_order_id}'
@@ -106,7 +105,8 @@ class AzureReservationsTests(ScenarioTest):
             'type': 'VirtualMachines',
             'location': 'westus'
         })
-        catalog = self.cmd('reservations catalog show --subscription-id {subscription} --reserved-resource-type {type} --location {location}').get_output_in_json()
+        catalog = self.cmd(
+            'reservations catalog show --subscription-id {subscription} --reserved-resource-type {type} --location {location}').get_output_in_json()
         self.assertGreater(len(catalog), 0)
         for entry in catalog:
             self.assertGreater(len(entry['terms']), 0)
@@ -116,10 +116,10 @@ class AzureReservationsTests(ScenarioTest):
 
     def test_update_reservation(self):
         self.kwargs.update({
-            'reservation_order_id': "0a47417c-cd30-4f67-add6-d631583e09f3",
-            'reservation_id': 'ae1fbdad-6333-4964-9f4c-83f7e2b7f44f',
+            'reservation_order_id': 'fe1341ea-4820-4ac9-9352-4136a6d8a252',
+            'reservation_id': '8e5963e2-000b-45bd-a1b4-305c9e5f89c9',
             'scope': '/subscriptions/d3ae48e5-dbb2-4618-afd4-fb1b8559cb80',
-            'instance_flexibility': "Off"
+            'instance_flexibility': 'Off'
         })
 
         single_reservation = self.cmd('reservations reservation update --reservation-order-id {reservation_order_id}'
@@ -134,8 +134,8 @@ class AzureReservationsTests(ScenarioTest):
 
     def test_split_and_merge(self):
         self.kwargs.update({
-            'reservation_order_id': "0a47417c-cd30-4f67-add6-d631583e09f3",
-            'reservation_id': 'ae1fbdad-6333-4964-9f4c-83f7e2b7f44f',
+            'reservation_order_id': '0af601f3-7868-44ee-b833-4d2e64ad3d70',
+            'reservation_id': '6dee7663-3e63-4115-aa4d-41e9a57f551e',
             'quantity1': 1,
             'quantity2': 2
         })
@@ -170,3 +170,49 @@ class AzureReservationsTests(ScenarioTest):
             self._validate_reservation(item)
             if 'Succeeded' in item['properties']['provisioningState']:
                 self.assertEqual(quantity_sum, item['properties']['quantity'])
+
+    def test_calculate_reservation_order(self):
+        self.kwargs.update({
+            'subid': 'd3ae48e5-dbb2-4618-afd4-fb1b8559cb80',
+            'sku': 'standard_b1ls',
+            'location': 'westus',
+            'reservedResourceType': 'VirtualMachines',
+            'term': 'P1Y',
+            'quantity': '2',
+            'displayName': 'test',
+            'appliedScopes': 'Shared',
+            'instanceFlexibility': 'Off',
+            'billingPlan': 'Monthly',
+            'appliedScopeType': 'Shared'
+        })
+        response = self.cmd('reservations reservation-order calculate --sku {sku} --location {location} --reserved-resource-type {reservedResourceType}'
+                            ' --billing-scope {subid} --term {term} --billing-plan {billingPlan} --display-name {displayName}'
+                            ' --quantity {quantity} --applied-scope-type {appliedScopeType}').get_output_in_json()
+        self.assertIsNotNone(response)
+        self.assertIsNotNone(response['properties']['reservationOrderId'])
+        self.assertEqual('standard_b1ls', response['properties']['skuDescription'])
+
+    def test_purchase_reservation_order(self):
+        self.kwargs.update({
+            'roid': 'd4ef7ec2-941c-4da7-8ec9-2f148255a0dc',
+            'subid': 'd3ae48e5-dbb2-4618-afd4-fb1b8559cb80',
+            'sku': 'standard_b1ls',
+            'location': 'westus',
+            'reservedResourceType': 'VirtualMachines',
+            'term': 'P1Y',
+            'quantity': '2',
+            'displayName': 'test',
+            'appliedScopes': 'Shared',
+            'instanceFlexibility': 'Off',
+            'billingPlan': 'Monthly',
+            'appliedScopeType': 'Shared'
+        })
+        response = self.cmd('reservations reservation-order purchase --reservation-order-id {roid} --sku {sku} --location {location} --reserved-resource-type {reservedResourceType}'
+                            ' --billing-scope {subid} --term {term} --billing-plan {billingPlan} --display-name {displayName}'
+                            ' --quantity {quantity} --applied-scope-type {appliedScopeType}').get_output_in_json()
+        self.assertIsNotNone(response)
+        self.assertGreater(response['etag'], 0)
+        self.assertIsNotNone(response['term'])
+        self.assertIsNotNone(response['billingPlan'])
+        self.assertIsNotNone(response['displayName'])
+        self.assertEqual(2, response['originalQuantity'])