+++ src/azure-cli/HISTORY.rst
@@ -82,6 +82,8 @@ Release History
 
 * Upgrading SDK Version to 0.6.0
 * Add billingplan details info after calling Get-Gatalogs
+* Add new command `az reservations reservation-order calculate` to calculate the price for a reservation
+* Add new command `az reservations reservation-order purchase` to purchase a new reservation
 
 **Rest**
 * `az rest` is now GA